+++ engine/bsql_engine/io/io.pyx
@@ -400,7 +400,8 @@ cpdef parseSchemaCaller(fileList, file_format_hint, args, extra_columns, ignore_
         return_object['types'].append(<underlying_type_t_type_id>(type))
     return_object['names'] = tableSchema.names
     return_object['calcite_to_file_indices']= tableSchema.calcite_to_file_indices
-    return_object['has_header_csv']= tableSchema.has_header_csv
+    return_object['has_header_csv'] = tableSchema.has_header_csv
+    return_object['row_groups_ids'] = tableSchema.row_groups_ids
 
     return return_object